refactor(elysia): nest search/knowledge/supersede — one file per endpoint
Per team-lead direction, each route module becomes a directory with:
  index.ts  — composes endpoints via .use()
  model.ts  — TypeBox schemas
  <verb>.ts — one file per endpoint (30-80 lines typical)

No behavior change — same handlers, same response shapes, same parse→500
remap in knowledge/learn.ts.
